Fix locmax issues
